Fix issue with mouse wheel stealing focus improperly.
As per GitHub Issue #1, when using the mouse wheel on the scroll bar, it became focused. There were other issues: using the mouse wheel over a window title caused the next click to be received by that window.

The cause was the same for all, and this patch aims at addressing them. Basically, click-and-drag should only work for the left mouse button, but it was not being button-specific.

Report by Szunti, fix by pgimeno.