Fix counting label in demo.
Upon closer examination, it's the demo's expectations that were wrong on what the value of gui.mousein should be. It's perfectly correct for it to report that the mouse is over the label, and the update function of the label failed to account for that, expecting that only the parent would be detected. Fixed now.

It worked just by chance when things were not done in a proper sequence (before 1ebb18c), because at the time the label's update function was called, gui.mousein was still not updated to point to the label.