Properly adjust width for elements and height for a text element.
On one hand, the width adjustment in element.setfont missed considering the standard margin (style.unit/2).

On the other hand, there was a mismatch in the calculation of the number of lines in a text element, caused by using a different width for font:getWrap and for printf when autosize was false. Since the previous call to element:setfont will now properly deal with the case when autosize is true, the whole block is conditioned on autosize being false.

There was a mismatch between the width used for