Properly check for invalid font in element.setfont().
'if pcall(...) then' will execute the "then" part only if there was no error, not if the nested function returned true, so the test would pass with any LÖVE Object instance. Changed the check to something lighter and that works in most cases (hopefully in all practical cases). It will still fail in either of these cases:

- setfont() is passed a userdata object which can't be indexed
- the function <passed object>.type() does not have the expected signature
- the function <passed object>.type() returns "Font" in unexpected cases

all of which look quite remote, and will hopefully be easy to debug should they arise.