Fix click position of text elements.
Text elements are drawn vertically centered, possibly above their defined initial position. This makes clicking them problematic, because they're visually out of their defined rectangle, resulting in the click being a bit off, e.g. with the checkbox label. With bigger fonts it's even worse.

If we modified the position to make the rectangle valid, we would depend on losslessly tracking the position changes according to font height, which might or might not be possible. Rather than going down that rabbit hole, it's easier to treat them as a special case when checking whether to change the mousein property (on which the clicked event depends), and that's what this change does, by transforming the mouse's coordinates to match the drawing pont.